+++ backend/src/users/users.service.ts
@@ -1,8 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './user.entity';
-import { RegisterUserDto } from './createUser.dto';
+import { RegisterUserDto } from 'src/auth/registerUser.dto';
 import { UpdateUserDto } from './updateUser.dto';
 
 @Injectable()
@@ -13,20 +13,47 @@ export class UsersService {
   ) {}
 
   async getAllUsers(): Promise<User[]> {
-    const allUsers: User[] = await this.usersRepository.find();
-    return allUsers;
+    return await this.usersRepository.find();
   }
 
-  async getUserById(id: number): Promise<User | null> {
-    return await this.usersRepository.findOneBy({ id });
+  async getUserById(id: number): Promise<User> {
+    if (isNaN(id)) {
+      throw new BadRequestException({ error: 'Invalid ID!' });
+    }
+
+    const user = await this.usersRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException({ error: 'User not found!' });
+    }
+    return user;
   }
 
-  async findUserByName(name: string): Promise<User | null> {
-    return await this.usersRepository.findOneBy({ name });
+  async getUserByIdWithRole(id: number): Promise<User> {
+    if (isNaN(id)) {
+      throw new BadRequestException({ error: 'Invalid ID!' });
+    }
+
+    const user = await this.usersRepository.findOne({ where: { id }, relations: ['role'] });
+    if (!user) {
+      throw new NotFoundException({ error: 'User not found!' });
+    }
+    return user;
   }
 
-  async findUserByEmail(email: string): Promise<User | null> {
-    return await this.usersRepository.findOneBy({ email });
+  async getUserByName(name: string): Promise<User> {
+    const user = await this.usersRepository.findOneBy({ name });
+    if (!user) {
+      throw new NotFoundException({ error: 'User not found!' });
+    }
+    return user;
+  }
+
+  async getUserByEmail(email: string): Promise<User> {
+    const user = await this.usersRepository.findOneBy({ email });
+    if (!user) {
+      throw new NotFoundException({ error: 'User not found!' });
+    }
+    return user;
   }
 
   async registerNewUser(user: RegisterUserDto): Promise<User> {
@@ -35,23 +62,49 @@ export class UsersService {
   }
 
   async deleteUserById(id: number): Promise<User> {
+    if (isNaN(id)) {
+      throw new BadRequestException({ error: 'Invalid ID!' });
+    }
+
     const user = await this.usersRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException({ error: 'User not found!' });
+    }
+
     await this.usersRepository.delete(id);
     return user;
   }
 
   async updateUserById(id: number, updateUserDto: UpdateUserDto): Promise<User> {
+    if (isNaN(id)) {
+      throw new BadRequestException({ error: 'Invalid ID!' });
+    }
+
     const user = await this.usersRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException({ error: 'User not found!' });
+    }
+
+    const isUpdated = await this.verifyChanges(user, updateUserDto);
+    if (!isUpdated) {
+      throw new BadRequestException({ error: 'No changes were made!' });
+    }
+
     Object.assign(user, updateUserDto);
     return await this.usersRepository.save(user);
   }
 
   async verifyChanges(user: User, updateUserDto: UpdateUserDto): Promise<boolean> {
     const { name, email } = user;
-    if (name !== updateUserDto.name) return true;
-    if (email !== updateUserDto.email) return true;
 
-    const isPasswordUpdated = await user.validatePassword(updateUserDto.password);
-    return !isPasswordUpdated;
+    if (updateUserDto.name && updateUserDto.name !== name) return true;
+    if (updateUserDto.email && updateUserDto.email !== email) return true;
+
+    if (updateUserDto.password) {
+      const isSamePassword = await user.validatePassword(updateUserDto.password);
+      if (!isSamePassword) return true;
+    }
+
+    return false;
   }
 }
